+++ packages/abstract-3d/src/renderers/svg/svg-geometries/svg-polygon.ts
@@ -1,4 +1,4 @@
-import { Polygon, vec3ZMean, Vec3, Vec2, vec3TransRot, vec3RotCombine, vec3Zero } from "../../../abstract-3d.js";
+import { Polygon, vec3ZMean, Vec3, Vec2, vec3TransRot, vec3RotCombine, vec3Zero, Material } from "../../../abstract-3d.js";
 import { gray, zElem, zOrderElement, transparent, SvgOptions } from "./shared.js";
 import { svgPolygon } from "../svg-encoding.js";
 import { rgbGrayScale } from "../../shared.js";
@@ -7,7 +7,7 @@ export function polygon(
   p: Polygon,
   point: (x: number, y: number) => Vec2,
   factor: number,
-  color: string,
+  material: Material,
   opts: SvgOptions,
   parentPos: Vec3,
   parentRot: Vec3
@@ -16,8 +16,9 @@ export function polygon(
   const rot = vec3RotCombine(parentRot, p.rot ?? vec3Zero);
   const rotatedPoints = p.points.map((p) => vec3TransRot(p, pos, rot));
   const points = rotatedPoints.map(({ x, y }) => point(x, y));
+  const color = material.normal;
   const [strokeColor, fill, strokeThickness] = opts.onlyStroke
     ? [opts.grayScale ? gray : color, opts.onlyStrokeFill, opts.stroke]
     : [transparent, opts.grayScale ? rgbGrayScale(color) : color, 0];
-  return [zElem(svgPolygon(factor, rot, points, fill, strokeColor, strokeThickness), vec3ZMean(...rotatedPoints))];
+  return [zElem(svgPolygon(factor, rot, points, fill, material.opacity ?? 1.0, strokeColor, strokeThickness), vec3ZMean(...rotatedPoints))];
 }