+++ packages/abstract-3d/src/renderers/svg/svg-geometries/svg-shape.ts
@@ -1,4 +1,4 @@
-import { vec3ZMean, Vec3, Vec2, vec3TransRot, vec3RotCombine, Shape, vec3, vec3Zero } from "../../../abstract-3d.js";
+import { vec3ZMean, Vec3, Vec2, vec3TransRot, vec3RotCombine, Shape, vec3, vec3Zero, Material } from "../../../abstract-3d.js";
 import { gray, zElem, zOrderElement, transparent, SvgOptions } from "./shared.js";
 import { svgPolygon } from "../svg-encoding.js";
 import { rgbGrayScale } from "../../shared.js";
@@ -7,7 +7,7 @@ export function shape(
   s: Shape,
   point: (x: number, y: number) => Vec2,
   factor: number,
-  color: string,
+  material: Material,
   opts: SvgOptions,
   parentPos: Vec3,
   parentRot: Vec3
@@ -16,8 +16,9 @@ export function shape(
   const rot = vec3RotCombine(parentRot, s.rot ?? vec3Zero);
   const rotatedPoints = s.points.map((p) => vec3TransRot(vec3(p.x, p.y, 0), pos, rot));
   const points = rotatedPoints.map(({ x, y }) => point(x, y));
+  const color = material.normal;
   const [strokeColor, fill] = opts.onlyStroke
     ? [opts.grayScale ? gray : color, opts.onlyStrokeFill]
     : [transparent, opts.grayScale ? rgbGrayScale(color) : color];
-  return [zElem(svgPolygon(factor, rot, points, fill, strokeColor, 0), vec3ZMean(...rotatedPoints))];
+  return [zElem(svgPolygon(factor, rot, points, fill, material.opacity ?? 1.0, strokeColor, 0), vec3ZMean(...rotatedPoints))];
 }